Substantial refactoring of contour file.
